linker: Consolidate WorkgroupSize constants
glslangValidator will usually both set an execution mode for an entry
point as well as defining a WorkgroupSize built-in with the same value.
That built-in is not directly attached to any entry point, and so after
linking multiple modules together one can end up with several of those
WorkgroupSize built-ins lying around with no way of knowing to which
entry point they should apply. Some drivers seem to rely on the
execution mode, but not all.

Before linking the modules together, we will try to remove those
constant WorkgroupSize and instead replace them with an execution mode
if there is not one already.